+++ src/Model/Offer/OfferAttributesPatch.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace DevLancer\VonHalsky\Model\Offer;
 
-use DevLancer\VonHalsky\Exception\InvalidRequestException;
 use DevLancer\VonHalsky\Model\RequestDtoInterface;
 use DevLancer\VonHalsky\Validation\RequestValidator;
 
@@ -14,9 +13,6 @@ final class OfferAttributesPatch implements RequestDtoInterface
     public function __construct(public readonly array $operations)
     {
         RequestValidator::offerAttributeOperations($operations);
-        if ($operations === []) {
-            throw new InvalidRequestException('Offer.attributes.operations', 'must not be empty');
-        }
     }
 
     public function jsonSerialize(): array